# Pin CacheEntryStatsCollector to fix performance bug