@ajkr 2021-06-12
Damn, right after submitting my review I remembered about MakeSharedCacheHandleGuard. Ignore my doubts, I agree it should entirely prevent edge cases of high collection frequency.